feat(front): add UI setting in the right menu panel, refacto(front): rename MenuLeftPanel* into Setting* because we use theses componenents outside the MenuLeftPanel
