+++ src/frontend/classic/components/SettingInput.svelte
