+++ src/frontend/classic/components/SettingItem.svelte
