+++ src/frontend/classic/components/SettingSelect.svelte
