+++ src/frontend/classic/components/SettingToggle.svelte
